app/requirements.txt
dash==1.0.0
dash-bootstrap-components==0.7.1
dash-core-components==1.0.0
dash-daq==0.1.5
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.0
numpy==1.16.4
pandas==0.24.2
beautifulsoup4==4.7.1
feather_format
